requirements.txt
boto3==1.42.85
botocore==1.42.85
jmespath==1.1.0
mysql-connector-python==9.6.0
numpy==2.4.4
pandas==3.0.2
psutil==7.2.2
py-cpuinfo==9.0.0
python-dateutil==2.9.0.post0
s3transfer==0.16.0
six==1.17.0
speedtest==0.0.1
urllib3==2.6.3
